Allow all separated lists in the language syntax to have a trailing separator.
This makes the language more uniform, but it also has a more practical benefit
in that it gives the programmer a way to signal to the formatter that the list
should be formatted on multiple lines. That way, the formatter will not have to
apply complex rules/heuristics to decide when a separated list should be on a
single line or multiple lines - it can just let the programmer be the arbiter
of good taste.

Also addressed some TODOs in the language parser by making tuple parsing use
ParseSeparatedList(). That method has also been simplified a fair bit since it
no longer has to handle as many cases.